reorganized components of menu to MenuPage & uniform import & export syntax of all components
