+++ frontend/src/App.jsx
@@ -1,29 +1,12 @@
 import { useState, useEffect } from 'react'; 
 import './App.css'; 
-import { CategoryList } from './components/CategoryList';
+import MenuPage from './pages/MenuPage'; 
 
 function App() {
 
-  const [menu, setMenu] = useState([])
-
-  useEffect(() => { 
-    const fetchMenuItems = async() => { 
-      try { 
-        const resp = await fetch('http://localhost:3000/menu'); 
-        const data = await resp.json(); // parse to json 
-        setMenu(data); 
-      } catch (error) { 
-        console.error('Error fetching meu items:', err)
-      }
-    }
-    fetchMenuItems(); 
-  }, [])
-
-
   return (
     <>
-     <h1>Chinatown Application</h1>
-     <CategoryList menu={menu} />
+     <MenuPage />
     </>
   )
 }
