+++ frontend/src/pages/MenuPage.jsx
@@ -1,0 +1,31 @@
+import { useState, useEffect } from 'react'; 
+// import './App.css'; 
+import CategoryList  from '../components/CategoryList';
+
+
+export default function MenuPage() {
+
+  const [menu, setMenu] = useState([])
+
+  useEffect(() => { 
+    const fetchMenuItems = async() => { 
+      try { 
+        const resp = await fetch('http://localhost:3000/menu'); 
+        const data = await resp.json(); // parse to json 
+        setMenu(data); 
+      } catch (error) { 
+        console.error('Error fetching meu items:', err)
+      }
+    }
+    fetchMenuItems(); 
+  }, [])
+
+
+  return (
+    <>
+     <h1>Chinatown Application</h1>
+     <CategoryList menu={menu} />
+    </>
+  )
+}
+
